Addressed review comments - Updated test function names - Moved integration tests to new folder /integration
